Cleanup to allow testing
