+++ hoprd/hoprd/Cargo.toml
@@ -91,6 +91,7 @@ hopr-async-runtime = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
+tokio = { workspace = true }
 
 [build-dependencies]
 vergen-gix = { version = "1.0.0", features = ["build"] }
