+++ hoprd/hoprd/src/lib.rs
@@ -89,3 +89,44 @@
 pub mod cli;
 pub mod config;
 pub mod errors;
+
+const LISTENING_SESSION_RETRANSMISSION_SERVER_PORT: u16 = 4677;
+
+#[derive(Debug, Clone)]
+pub struct HoprServerReactor {}
+
+#[hopr_lib::async_trait]
+impl hopr_lib::HoprSessionServerActionable for HoprServerReactor {
+    #[tracing::instrument(level = "debug", skip(self, session))]
+    async fn process(&self, session: hopr_lib::HoprSession) -> hopr_lib::errors::Result<()> {
+        let server_port = LISTENING_SESSION_RETRANSMISSION_SERVER_PORT;
+
+        tracing::debug!("Creating a connection to the TCP server on port 127.0.0.1:{server_port}...");
+        let mut tcp_bridge = tokio::net::TcpStream::connect(std::net::SocketAddr::new(
+            std::net::IpAddr::V4(std::net::Ipv4Addr::new(127, 0, 0, 1)),
+            server_port,
+        ))
+        .await
+        .map_err(|e| {
+            hopr_lib::errors::HoprLibError::GeneralError(format!(
+                "Could not bridge the incoming session to port {server_port}: {e}"
+            ))
+        })?;
+
+        tcp_bridge.set_nodelay(true).map_err(|e| {
+            hopr_lib::errors::HoprLibError::GeneralError(format!(
+                "Could not set the TCP_NODELAY option for the bridged session to port {server_port}: {e}"
+            ))
+        })?;
+
+        tracing::debug!("Bridging the session to the TCP server...");
+        tokio::task::spawn(async move {
+            match tokio::io::copy_bidirectional(&mut tokio_util::compat::FuturesAsyncReadCompatExt::compat(session), &mut tcp_bridge).await {
+                Ok(bound_stream_finished) => tracing::info!("Server bridged session through TCP port {server_port} ended with {bound_stream_finished:?} bytes transferred in both directions."),
+                Err(e) => tracing::error!("The TCP server stream (port {server_port}) is closed: {e:?}")
+            }
+        });
+
+        Ok(())
+    }
+}
